Removes macros from system.cpp for readability...

+++ pulsar/exception/PulsarException.cpp
@@ -3,7 +3,7 @@
  */
 
 #include "pulsar/exception/PulsarException.hpp"
-
+#include "pulsar/math/Cast.hpp"
 #include <iomanip>
 
 namespace pulsar {
@@ -35,7 +35,7 @@ void PulsarException::build_str_(std::string & outstr,
     // Get the first line of the formatted output and
     // output it to the stream
     std::getline(sstr, str);
-    ss << std::setw(24) << key << std::setw(oldw) << " : " << str;
+    ss << std::setw(24) << key << std::setw(numeric_cast<int>(oldw)) << " : " << str;
 
     // For the rest of the lines, output the values aligned
     // with the first
